+++ jupyter_cassini_server/handlers.py
@@ -1,21 +1,19 @@
-import functools
-from typing import TypeVar, Type, Callable, Any, Dict, Literal, Union, List, cast, Set
+from typing import TypeVar, Callable, Any, Union
 import datetime
-import urllib.parse
 
 from jupyter_server.base.handlers import APIHandler
 from jupyter_server.utils import url_path_join
 
 import tornado
-from pydantic import BaseModel
 
 from cassini import env
-from cassini.core import TierABC, NotebookTierBase
+from cassini.core import NotebookTierBase
+
+from jupyter_cassini_server.safety import needs_project, with_types
+from jupyter_cassini_server.serialisation import serialize_branch
 
 from .schema.models import (
     NewChildInfo,
-    ChildClsInfo,
-    TreeChildResponse,
     TreeResponse,
     TierInfo,
     LookupGetParametersQuery,
@@ -24,205 +22,6 @@ from .schema.models import (
     Status,
     Status1,
 )
-
-Q = TypeVar("Q", bound=BaseModel)
-R = TypeVar("R", bound=BaseModel)
-S = TypeVar("S", bound=APIHandler)
-
-DecoratoredType = Callable[[APIHandler, Q], R]
-MethodType = Callable[[APIHandler], None]
-
-
-def with_types(
-    query_model: Type[Q],
-    response_model: Type[R],
-    method: Union[Literal["GET"], Literal["POST"]],
-) -> Callable[[Callable[[S, Q], R]], Callable[[S], None]]:
-
-    def wrapper(func: Callable[[S, Q], R]) -> Callable[[S], None]:
-
-        if method == "GET":
-
-            def wrap_get(self: S) -> None:
-                query: Dict[str, Union[str, List[str]]] = {}
-
-                raw_arguments = urllib.parse.parse_qs(self.request.query)
-
-                for key, raw_val in raw_arguments.items():
-                    val = "".join(raw_val)
-
-                    if (
-                        "[]" in key
-                    ):  # we aren't compatible with object encoding and rely on form no explode.
-                        query[key] = val.split(",")
-                    else:
-                        query[key] = val
-                try:
-                    response = response_model.model_validate(
-                        func(self, query_model.model_validate(query))
-                    )
-                    self.finish(response.model_dump_json())
-                except ValueError:
-                    self.send_error(404)
-
-            return wrap_get
-        elif method == "POST":
-
-            def wrap_post(self: S) -> None:
-                try:
-                    query = self.get_json_body()
-                    response = response_model.model_validate(
-                        func(self, query_model.model_validate(query))
-                    )
-                    self.finish(response.model_dump_json())
-                except ValueError:
-                    self.send_error(404)
-
-            return wrap_post
-        else:
-            raise NotImplementedError
-
-    return wrapper
-
-
-F = TypeVar("F", bound=Callable[..., Any])
-
-
-def needs_project(meth: F) -> F:
-
-    @functools.wraps(meth)
-    def wraps(self, *args, **kwargs):
-        if not env.project:
-            self.finish(
-                "Current project not set, jupyterlab needs to be launched by Cassini"
-            )
-            return
-
-        return meth(self, *args, **kwargs)
-
-    return cast(F, wraps)
-
-
-def serialize_child(tier: TierABC) -> TreeChildResponse:
-    """
-    Note, doesn't populate children field... maybe will later...
-    """
-    assert env.project
-    project_folder = env.project.project_folder
-
-    if isinstance(tier, NotebookTierBase):
-        notebookPath = tier.file.relative_to(project_folder).as_posix()
-
-        metaPath = tier.meta_file.relative_to(project_folder).as_posix()
-        additionalMeta = {
-            key: tier.meta.get(key)
-            for key in tier.meta.keys()
-            if key not in ["description", "conclusion", "started"]
-        }
-
-        started = tier.started.astimezone().isoformat()
-
-        if tier.description:
-            info = tier.description.split("\n")[0]
-        else:
-            info = None
-
-        if tier.conclusion:
-            outcome = tier.conclusion.split("\n")[0]
-        else:
-            outcome = None
-
-        if tier.highlights_file:
-            hltsPath = tier.highlights_file.relative_to(project_folder).as_posix()
-        else:
-            hltsPath = None
-    else:
-        notebookPath = None
-        info = None
-        outcome = None
-        started = None
-        metaPath = None
-        hltsPath = None
-        additionalMeta = {}
-
-    return TreeChildResponse(
-        name=tier.name,
-        info=info,
-        outcome=outcome,
-        started=started,
-        metaPath=metaPath,
-        hltsPath=hltsPath,
-        notebookPath=notebookPath,
-        additionalMeta=additionalMeta,
-    )
-
-
-def serialize_branch(tier: TierABC) -> TreeResponse:
-    assert env.project
-
-    core = serialize_child(tier)
-    folder = tier.folder.relative_to(env.project.project_folder).as_posix()
-
-    child_cls = tier.child_cls
-
-    if not child_cls:
-        return TreeResponse(
-            name=core.name,
-            outcome=core.outcome,
-            started=core.started,
-            hltsPath=core.hltsPath,
-            metaPath=core.metaPath,
-            notebookPath=core.notebookPath,
-            additionalMeta=core.additionalMeta,
-            folder=folder,
-            childClsInfo=None,
-            children={},
-        )
-
-    child_metas: Set[str] = set()
-    children = {}
-
-    for child in tier:
-        children[child.id] = serialize_child(child)
-
-        if isinstance(child, NotebookTierBase):
-            child_metas.update(child.meta.keys())
-
-    if issubclass(child_cls, NotebookTierBase):
-        child_metas.discard("name")
-        child_metas.discard("started")
-        child_metas.discard("description")
-        child_metas.discard("conclusion")
-
-        child_metaNames = list(child_metas)
-
-        child_templates = [
-            template.name for template in child_cls.get_templates(env.project)
-        ]
-    else:
-        child_templates = []
-        child_metaNames = []
-
-    child_cls_info = ChildClsInfo(
-        name=child_cls.pretty_type,
-        idRegex=child_cls.id_regex,
-        namePartTemplate=child_cls.name_part_template,
-        templates=child_templates,
-        metaNames=child_metaNames,
-    )
-
-    return TreeResponse(
-        name=core.name,
-        outcome=core.outcome,
-        started=core.started,
-        hltsPath=core.hltsPath,
-        metaPath=core.metaPath,
-        notebookPath=core.notebookPath,
-        additionalMeta=core.additionalMeta,
-        folder=folder,
-        childClsInfo=child_cls_info,
-        children=children,
-    )
 
 
 class LookupHandler(APIHandler):
